Seaparando logica
