--- src/main/scala/fiuba/fp/Logger.scala
@@ -1,5 +1,0 @@
-package fiuba.fp
-
-class Logger {
-
-}
